+++ tests/test_query.py
@@ -1,14 +1,11 @@
 from datetime import datetime
+
 import taos
-import pytest
 
-@pytest.fixture
-def conn():
-    return taos.connect()
 
-def test_query(conn):
-    # type: (taos.TaosConnection) -> None
+def test_query():
     """This test will use fetch_block for rows fetching, significantly faster than rows_iter"""
+    conn = taos.connect()
     conn.execute("drop database if exists test_query_py")
     conn.execute("create database if not exists test_query_py")
     conn.execute("use test_query_py")
@@ -20,10 +17,21 @@ def test_query(conn):
     fields = result.fields
     for field in fields:
         print("field: %s" % field)
+
+    # test re-consume fields
+    flag = 0
+    for _ in fields:
+        flag += 1
+    assert flag == 3
+
     start = datetime.now()
     for row in result:
         print(row)
         None
+
+    for row in result.rows_iter():
+        print(row)
+    
     result = conn.query("select * from tb1 limit 1")
     results = result.fetch_all_into_dict()
     print(results)
@@ -34,22 +42,5 @@ def test_query(conn):
     result.close()
     conn.close()
 
-def _test_query_row_iter(conn):
-    """This test will use fetch_row for each row fetching, this is the only way in async callback"""
-    result = conn.query("select * from log.log limit 10000")
-    fields = result.fields
-    for field in fields:
-        print("field: %s" % field)
-    start = datetime.now()
-    for row in result.rows_iter():
-        # print(row)
-        None
-    end = datetime.now()
-    elapsed = end - start
-    print("elapsed time: ", elapsed)
-    result.close()
-    conn.close()
-
 if __name__ == "__main__":
-    conn1 = taos.connect()
-    test_query(conn1)
+    test_query()